Calendar Checking, Creating, and Deleting
Implemented code that checks if the user has a Tic Talk calendar or not.
If the user does not have a calendar yet, the user is prompted with a
modal that asks them if creating a calendar is okay. Upon pressing a
buton in the modal, the Tic Talk calendar is created. Also added a
method to delete the Tic Talk calendar for testing purposes.